file-psd: fix integer overflow when reading path anchor points
The variable 'cntr' represents the index of all control points of a Photoshop path. It was declared too small, which caused problems with detailed path objects. 'num_rec' is of type 'gint16', but the number of control points is 'num_rec * 6', so 'cntr' must be able to cover that range

(cherry picked from commit a8d1135110bed67cd46f794d7a14fb22ca4f9777)

+++ plug-ins/file-psd/psd-image-res-load.c
@@ -1538,7 +1538,7 @@ load_resource_2000 (const PSDimageres  *res_a,
   gint16     init_fill = 0;
   gint16     num_rec;
   gint16     path_rec;
-  gint16     cntr;
+  gint       cntr;
   gint       image_width;
   gint       image_height;
   gint       i;
